templates: Convert mail settings pages
This one is rather tricky. We do a major overhaul of URLs and general
flow of confirmations, relying heavily on the messages framework to
avoid the need to have separate pages. Previously, configuring opt-in or
opt-out of email involved the following:

  GET /mail/
    POST /mail/
      [some validation]
      POST /mail/optout/  # or optin

If the last step threw an error, we'd stay on '/mail/optout/' or
'/mail/optin/' with a displayed error and tell the user to correct the
mistake. Now, we simply do this:

  GET /mail/
    POST /mail/
      [some validation]
      GET /mail/<email>/
      POST /mail/<email>/

Error messages are propagated via the messages framework with all
non-field validation errors resulting in a redirect to the homepage.

Signed-off-by: Stephen Finucane <stephen@that.guru>