parser: remove in-reply-to/references comments
When contributors using Gnus reply to a patch, the In-Reply-To &
References header fields include comments in the form described in
remove_rfc2822_comments spec. These comments can lead to threading
issues where users' replies don't get associated to the patch they
reply to because the added comment will be part of the message-id when
parsed. Following RFC2822, a comment is not supposed to affect threading
in any way, so we remove it. To use regex, we assume that comments are
not nested; further changes to more thoroughly match the RFC2822 grammar
remains for anyone interested.

Signed-off-by: Raxel Gutierrez <raxel@google.com>
Closes: #399