migrations: don't go to the db for 0041_python3 migration
When we moved to Python 3, makemigrations wanted to change a bunch of
things, see commit 4ad87ed72aec ("migrations: Add the Python 3 patch").

The change is, as observed then, just to make django happy; it's not
supposed to change the database at all.

So make the migration change the state as seen by Django only.
This makes the migration ~instant, even for a huge database.

Signed-off-by: Daniel Axtens <dja@axtens.net>
Reviewed-by: Stephen Finucane <stephen@that.guru>
Related: #420