docker: Bump tested database versions
We were attempting to test with PostgreSQL 9.6, but Django 4.0 requires
PostgreSQL 10.0 or greater [1]. This was causing failures like those
seens at [2]. Configure jobs to use 'latest' like we do in CI to avoid
this issue.

While we're here, we also switch from MariaDB to MySQL and use the
'latest' tag of this. Once again, this was what we were actually using
in CI and doing this limits future work.

[1] https://docs.djangoproject.com/en/4.0/ref/databases/#postgresql-notes
[2] https://stackoverflow.com/q/69477858

Signed-off-by: Stephen Finucane <stephen@that.guru>