parser: Ignore CFWS in In-Reply-To, References headers
RFC2822 states that [1] a comment or folding white space is permitted to
be inserted before or after a msg-id in in the Message-ID, In-Reply-To
or References fields. Allow for this.

[1] https://tools.ietf.org/html/rfc2822#section-3.6.4

Conflicts:
  patchwork/tests/test_parser.py

NOTE(stephenfin): Conflicts are due to the absence of commits f5cd52144c
("parser: Add 'X-Patchwork-Action-Required' header") and 94d75a1c78
("Blackify code"), neither of which we want to backport.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Reviewed-by: DJ Delorie <dj@redhat.com>
Closes: #399