parser: Ignore CFWS in Message-ID header
We recently started stripping comments and folding white space from the
In-Reply-To and References headers. Do so also for the Message-ID
header. Because of the importance of the Message-ID header, we accept
even non-compliant headers and because we now have a pattern for this,
we also start (re-)accepting non-compliant In-Reply-To headers.

Conflicts:
  patchwork/parser.py

NOTE(stephenfin): Conflicts are due to the absence of commit 94d75a1c78
("Blackify code") which we don't want to backport.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Related: #399