requirements: Bump Django to 3.2.x, djangorestframework to 4.16.0
There are two issues to be addressed:

  RemovedInDjango50Warning: Passing response to assertFormError() is
  deprecated. Use the form object directly:

  RemovedInDjango50Warning: The "default.html" templates for forms and
  formsets will be removed. These were proxies to the equivalent
  "table.html" templates, but the new "div.html" templates will be the
  default from Django 5.0. Transitional renderers are provided to allow
  you to opt-in to the new output style now. See
  https://docs.djangoproject.com/en/4.1/releases/4.1/ for more details

Nothing complicated in fixing either of these. For the former, we must
do as we're told and use the form object directly. For the latter, we
need to configure our own form renderer so we can continue using the
table form renderer for now.

Signed-off-by: Stephen Finucane <stephen@that.guru>