Add series list view
Add a series list view - as opposed to the default view which shows
patches, this new view shows a list of all series. This is
a much easier way to navigate and manage large repositories with 100+
patches (Some series alone contain more than 20-30 patches in a single
series). This view will allow simple management of patch series.
This version is MVP - and contains the basic view needed to navigate
around.