templates/submission: Fix alignment of commit message
Preformatted content must not be indented
because any space is kept in the output,
making the content wrongly indented.

When aligning message headers to the left,
the new HTML code has been indented
including some preformatted content indented with two spaces.

The fix is to remove the indent of the content.

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Fixes: fe34ab2ffad3 ("patch-detail: left align message headers")
Reviewed-by: Robin Jarry <robin@jarry.cc>
Reviewed-by: Stephen Finucane <stephen@that.guru>