views: Don't show chevron if user can't change sort order
We prevent users changing the column that sorting is done on when
allowing users to change the order of patches in a bundle. However, we
still show a chevron and clicking the link will appear to do something.
This is confusing/misleading. Remove the chevron and the link in this
situation.

Signed-off-by: Stephen Finucane <stephen@that.guru>